Moved the DuplicateWarning to duplicates folder
